Added Input dat functions from scale operator to common utils
