+++ tests/functional-tests/tests/operator_test.py
@@ -6,6 +6,7 @@ import pytest
 from kubernetes import client
 from kubernetes.client.rest import ApiException
 import ibm_spectrum_scale_csi.scale_operator as scaleop
+import ibm_spectrum_scale_csi.common_utils.input_data_functions as inputfunc
 LOGGER = logging.getLogger()
 pytestmark = pytest.mark.csioperator
 
@@ -13,11 +14,11 @@ pytestmark = pytest.mark.csioperator
 def _values(request):
 
     global kubeconfig_value, clusterconfig_value, namespace_value
-    kubeconfig_value, clusterconfig_value, operator_namespace, test_namespace, _, operator_yaml = scaleop.get_cmd_values(request)
+    kubeconfig_value, clusterconfig_value, operator_namespace, test_namespace, _, operator_yaml = inputfunc.get_cmd_values(request)
     namespace_value = operator_namespace
     condition = scaleop.check_ns_exists(kubeconfig_value, namespace_value)
     operator = scaleop.Scaleoperator(kubeconfig_value, namespace_value, operator_yaml)
-    read_file = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    read_file = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     scaleop.filesetfunc.cred_check(read_file)
     fileset_exist = scaleop.filesetfunc.fileset_exists(read_file)
     operator.create()
@@ -36,7 +37,7 @@ def _values(request):
 
 
 def test_get_version(_values):
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     scaleop.filesetfunc.get_scale_version(test)
     scaleop.get_kubernetes_version(kubeconfig_value)
     scaleop.csioperatorfunc.get_operator_image()
@@ -46,7 +47,7 @@ def test_operator_deploy(_values):
 
     LOGGER.info("test_operator_deploy")
     LOGGER.info("Every input is correct should run without any error")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
     if operator_object.check() is True:
@@ -71,7 +72,7 @@ def test_operator_deploy(_values):
 
 def test_wrong_cluster_id(_values):
     LOGGER.info("test_wrong_cluster_id : cluster ID is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     wrong_id = str(random.randint(0, 999999999999999999))
 
     for cluster in test["custom_object_body"]["spec"]["clusters"]:
@@ -106,7 +107,7 @@ def test_wrong_cluster_id(_values):
 
 def test_wrong_primaryFS(_values):
     LOGGER.info("test_wrong_primaryFS : primaryFS is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     wrong_primaryFs = scaleop.csiobjectfunc.randomStringDigits()
 
     for cluster in test["custom_object_body"]["spec"]["clusters"]:
@@ -141,7 +142,7 @@ def test_wrong_primaryFS(_values):
 
 def test_wrong_guihost(_values):
     LOGGER.info("test_wrong_guihost : gui host is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     wrong_guiHost = scaleop.csiobjectfunc.randomStringDigits()
     test["guiHost"] = wrong_guiHost
     for cluster in test["custom_object_body"]["spec"]["clusters"]:
@@ -178,7 +179,7 @@ def test_wrong_guihost(_values):
 
 def test_wrong_gui_username(_values):
     LOGGER.info("test_wrong_gui_username : gui username is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     test["username"] = scaleop.csiobjectfunc.randomStringDigits()
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
@@ -206,7 +207,7 @@ def test_wrong_gui_username(_values):
 
 def test_wrong_gui_password(_values):
     LOGGER.info("test_wrong_gui_password : gui password is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     test["password"] = scaleop.csiobjectfunc.randomStringDigits()
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
@@ -243,7 +244,7 @@ def test_wrong_gui_password(_values):
 
 def test_wrong_secret_object_name(_values):
     LOGGER.info("test_wrong_secret_object_name : secret object name is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     secret_name_wrong = scaleop.csiobjectfunc.randomString()
 
     for cluster in test["custom_object_body"]["spec"]["clusters"]:
@@ -258,7 +259,7 @@ def test_wrong_secret_object_name(_values):
 
 def test_random_gpfs_primaryFset_name(_values):
     LOGGER.info("test_random_gpfs_primaryFset_name : gpfs primary Fset name is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     random_primaryFset = scaleop.csiobjectfunc.randomStringDigits()
     test["primaryFset"] = random_primaryFset
     for cluster in test["custom_object_body"]["spec"]["clusters"]:
@@ -298,7 +299,7 @@ def test_random_gpfs_primaryFset_name(_values):
 def test_secureSslMode(_values):
     LOGGER.info("test_secureSslMode")
     LOGGER.info("secureSslMode is True while cacert is not available")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
 
     for cluster in test["custom_object_body"]["spec"]["clusters"]:
         if "primary" in cluster.keys():
@@ -343,7 +344,7 @@ Removing this testcase as scaleHostpath is no longer needed
 def test_wrong_gpfs_filesystem_mount_point(_values):
     LOGGER.info("test_wrong_gpfs_filesystem_mount_point")
     LOGGER.info("gpfs filesystem mount point is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     wrong_scaleHostpath = scaleop.csiobjectfunc.randomStringDigits()
     test["custom_object_body"]["spec"]["scaleHostpath"] = wrong_scaleHostpath
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -379,7 +380,7 @@ def test_wrong_gpfs_filesystem_mount_point(_values):
 def test_unlinked_primaryFset(_values):
     LOGGER.info("test_unlinked_primaryFset")
     LOGGER.info("unlinked primaryFset expected : object created successfully")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     if(not(scaleop.filesetfunc.fileset_exists(test))):
         scaleop.filesetfunc.create_fileset(test)
     scaleop.filesetfunc.unlink_fileset(test)
@@ -411,7 +412,7 @@ def test_existing_primaryFset(_values):
     LOGGER.info("test_existing_primaryFset")
     LOGGER.info(
         "linked existing primaryFset expected : object created successfully")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     if(not(scaleop.filesetfunc.fileset_exists(test))):
         scaleop.filesetfunc.create_fileset(test)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -441,7 +442,7 @@ def test_unmounted_primaryFS(_values):
     LOGGER.info("test_unmounted_primaryFS")
     LOGGER.info(
         "primaryFS is unmounted and expected : custom object should give error")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     scaleop.filesetfunc.unmount_fs(test)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
@@ -480,7 +481,7 @@ def test_unmounted_primaryFS(_values):
 def test_non_deafult_attacher(_values):
     LOGGER.info("test_non_deafult_attacher")
     LOGGER.info("attacher image name is changed")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     deployment_attacher_image = "quay.io/k8scsi/csi-attacher:v1.2.1"
     test["custom_object_body"]["spec"]["attacher"] = deployment_attacher_image
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -511,7 +512,7 @@ def test_non_deafult_attacher(_values):
 def test_non_deafult_provisioner(_values):
     LOGGER.info("test_non_deafult_provisioner")
     LOGGER.info("provisioner image name is changed")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     deployment_provisioner_image = "quay.io/k8scsi/csi-provisioner:v1.6.0"
     test["custom_object_body"]["spec"]["provisioner"] = deployment_provisioner_image
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -542,7 +543,7 @@ def test_non_deafult_provisioner(_values):
 def test_correct_cacert(_values):
     LOGGER.info("test_secureSslMode with correct cacert file")
     LOGGER.info("correct cacert file is given")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
 
     if not("local_cacert_name" in test):
         test["local_cacert_name"] = "test-cacert-configmap"
@@ -584,7 +585,7 @@ def test_correct_cacert(_values):
 def test_cacert_with_secureSslMode_false(_values):
     LOGGER.info("test_cacert_with_secureSslMode_false")
     LOGGER.info("secureSslMode is false with correct cacert file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
 
     if not("local_cacert_name" in test):
         test["local_cacert_name"] = "test-cacert-configmap"
@@ -626,7 +627,7 @@ def test_cacert_with_secureSslMode_false(_values):
 def test_wrong_cacert(_values):
     LOGGER.info("secureSslMode true with wrong cacert file")
     LOGGER.info("test_wrong_cacert")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
 
     if not("local_cacert_name" in test):
         test["local_cacert_name"] = "test-cacert-configmap"
@@ -680,7 +681,7 @@ def test_wrong_cacert(_values):
 def test_nodeMapping(_values):
     LOGGER.info("test_nodeMapping")
     LOGGER.info("nodeMapping is added to the cr file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     LOGGER.debug(test)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
 
@@ -714,7 +715,7 @@ def test_nodeMapping(_values):
 def test_attacherNodeSelector(_values):
     LOGGER.info("test_attacherNodeSelector")
     LOGGER.info("attacherNodeSelector is added to the cr file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
     if operator_object.check() is True:
@@ -750,7 +751,7 @@ def test_attacherNodeSelector(_values):
 def test_provisionerNodeSelector(_values):
     LOGGER.info("test_provisionerNodeSelector")
     LOGGER.info("provisionerNodeSelector is added to the cr file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
     if operator_object.check() is True:
@@ -786,7 +787,7 @@ def test_provisionerNodeSelector(_values):
 def test_pluginNodeSelector(_values):
     LOGGER.info("test_pluginNodeSelector")
     LOGGER.info("pluginNodeSelector is added to the cr file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
     if operator_object.check() is True:
@@ -822,7 +823,7 @@ def test_pluginNodeSelector(_values):
 def test_non_deafult_snapshotter(_values):
     LOGGER.info("test_non_deafult_snapshotter")
     LOGGER.info("snapshotter image name is changed")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     deployment_snapshotter_image = "us.gcr.io/k8s-artifacts-prod/sig-storage/csi-snapshotter:v4.1.1"
     test["custom_object_body"]["spec"]["snapshotter"] = deployment_snapshotter_image
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -853,7 +854,7 @@ def test_non_deafult_snapshotter(_values):
 def test_non_deafult_livenessprobe(_values):
     LOGGER.info("test_non_deafult_livenessprobe")
     LOGGER.info("livenessprobe image name is changed")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     deployment_livenessprobe_image = "us.gcr.io/k8s-artifacts-prod/sig-storage/livenessprobe:v2.3.0"
     test["custom_object_body"]["spec"]["livenessprobe"] = deployment_livenessprobe_image
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -885,7 +886,7 @@ def test_non_deafult_livenessprobe(_values):
 def test_non_deafult_resizer(_values):
     LOGGER.info("test_non_deafult_resizer")
     LOGGER.info("resizer image name is changed")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     deployment_resizer_image = "us.gcr.io/k8s-artifacts-prod/sig-storage/csi-resizer:v1.3.0"
     test["custom_object_body"]["spec"]["resizer"] = deployment_resizer_image
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
@@ -916,7 +917,7 @@ def test_non_deafult_resizer(_values):
 def test_snapshotterNodeSelector(_values):
     LOGGER.info("test_snapshotterNodeSelector")
     LOGGER.info("snapshotterNodeSelector is added to the cr file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
     if operator_object.check() is True:
@@ -952,7 +953,7 @@ def test_snapshotterNodeSelector(_values):
 def test_resizerNodeSelector(_values):
     LOGGER.info("test_resizerNodeSelector")
     LOGGER.info("resizerNodeSelector is added to the cr file")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
     operator_object = scaleop.Scaleoperatorobject(test, kubeconfig_value)
     operator_object.create()
     if operator_object.check() is True:
@@ -987,7 +988,7 @@ def test_resizerNodeSelector(_values):
 
 def test_wrong_kubeletRootDirPath(_values):
     LOGGER.info("test_wrong_kubeletRootDirPath : kubeletRootDirPath is wrong")
-    test = scaleop.read_operator_data(clusterconfig_value, namespace_value)
+    test = inputfunc.read_operator_data(clusterconfig_value, namespace_value)
 
     test["custom_object_body"]["spec"]["kubeletRootDirPath"] = f"/{scaleop.csiobjectfunc.randomString()}/{scaleop.csiobjectfunc.randomString()}"
